refactor(rn-parser): rn-parser rename
