+++ src/reactNative.parser/parsers/examples.parser.ts
@@ -1,4 +1,4 @@
-import { CO } from '../kitten.parser.options';
+import { CO } from '../reactNative.parser.options';
 import { Sample } from '../../model/class/example';
 
 export class ExamplesParser {
