+++ src/reactNative.parser/parsers/hocParams.parser.ts
@@ -1,5 +1,5 @@
 import { Param } from '../../model';
-import { CO } from '../kitten.parser.options';
+import { CO } from '../reactNative.parser.options';
 
 export class HocParamsParser {
