+++ src/reactNative.parser/parsers/types.parser.ts
@@ -1,5 +1,5 @@
 import { Type } from '../../model';
-import { CO } from '../kitten.parser.options';
+import { CO } from '../reactNative.parser.options';
 
 export class TypesParser {
