+++ src/reactNative.parser/reactNative.parser.options.ts
